+++ config/splat.us.dra.yaml
@@ -181,6 +181,7 @@ segments:
       - [0x47BB8, c, 47BB8]
       - [0x4F45C, c, 4F45C]
       - [0x52860, c, 52860]
+      - [0x5298C, c, 5298C]
       - [0x5D6C4, c, 5D6C4]
       - [0x627C4, c, 627C4]
       - [0x62D70, c, 62D70]
