+++ include/game.h
@@ -149,6 +149,7 @@ typedef struct {
     /* 0x1A */ s16 unk1A;
     /* 0x1C */ s16 unk1C;
 } MenuContext; // size = 0x1E
+#define SIZEOF_MENUCONTEXT (0x1E)
 
 typedef struct {
     /* 0x0 */ u8 tileLayoutId;
@@ -911,6 +912,13 @@ typedef struct {
     /* 80073104 */ u32 flags;
 } BgLayer;
 
+typedef struct {
+    u32 rightHand;
+    u32 leftHand;
+    u32 jump;
+    u32 dodge;
+} ConfigButtons;
+
 extern s32 D_8003925C;
 extern bool g_IsTimeAttackUnlocked;
 
@@ -939,12 +947,15 @@ extern GameApi g_api;
 extern s32 D_8003C8B8;
 extern u32 D_8003C8C4;
 extern s32 g_roomCount;
+extern s32 g_blinkTimer;
 extern s32 D_8003C99C;
-extern u16 D_8003C9C8;
-extern s16 D_8003C9CC[];
 extern s32 g_CurrentPlayableCharacter;
 extern s32 D_8003C9A4; // related to when player change stages?
-extern s32 g_blinkTimer;
+extern u16 D_8003C9C8;
+extern s16 D_8003C9CC[];
+extern s32 g_menuButtonSettingsCursorPos;
+extern ConfigButtons g_menuButtonSettingsConfig;
+extern s32 D_8003C9FC;
 extern s32 g_menuMainCursorIndex;
 extern s32 g_menuRelicsCursorIndex[];
 extern s32 g_SettingsCloakMode;
