+++ src/dra/dra.h
@@ -28,6 +28,10 @@ extern u16 g_saveIconPalette[0x10][0x10];
 extern u8* g_saveIconTexture[0x10];
 extern s32 D_800A2438;
 extern RoomBossTeleport D_800A297C[];
+extern s32 D_800A2D68;
+extern s32 D_800A2D6C;
+extern u8 c_chPlaystationButtons[];
+extern u8 c_chShoulderButtons[];
 extern Unkstruct_800A2D98 D_800A2D98[];
 extern u8 D_800A2EE8[];
 extern u8 D_800A2EED;
@@ -55,7 +59,7 @@ extern u32 g_playerEquip[];
 extern RoomTeleport D_800A245C[];
 extern s32 D_800A2464[]; // D_800A245C[0].stageId
 extern const char* c_strALUCARD;
-extern const char* c_strSTR;
+extern const char** c_strSTR;
 extern const char* c_strCON;
 extern const char* c_strINT;
 extern const char* c_strLCK;
@@ -90,6 +94,7 @@ extern const char* c_strMono;
 extern const char* c_strWindow;
 extern const char* c_strTime;
 extern const char* c_strALUCART;
+extern const char* D_800A83AC[];
 extern const char* c_strSSword;
 extern s32 D_800A3194[];
 extern Lba D_800A3C40[]; // g_lba
@@ -97,6 +102,7 @@ extern Unsktruct_800EAF28* D_800A3B5C[];
 extern SubweaponDef g_Subweapons[];
 extern SpellDef g_SpellDefs[];
 extern EnemyDef g_EnemyDefs[];
+extern s32 c_arrExpNext[];
 extern Equipment D_800A4B04[];
 extern Unkstruct_800A4B12 D_800A4B1D[];
 extern Accessory D_800A7718[];
@@ -441,7 +447,7 @@ s32 func_800F62E8(s32 arg0);
 void func_800FF7B8(s32 arg0);
 void func_800F98AC(s32 arg0, s32 arg1);
 void func_800F99B8(s32 arg0, s32 arg1, s32 arg2);
-void DrawMenuChar(char ch, int x, int y, MenuContext* context);
+void DrawMenuChar(u8 ch, int x, int y, MenuContext* context);
 void DrawMenuStr(const char* str, s32 x, s32 y, MenuContext* context);
 void DrawMenuInt(s32 value, s32 x, s32 y, MenuContext*);
 void DrawSettingsReverseCloak(MenuContext* context);
